requirements.txt
altgraph==0.16.1
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.6.3
bing-image-downloader==1.0.4
bs4==0.0.1
cached-property==1.5.2
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.7
configparser==5.0.2
crayons==0.4.0
dictdiffer==0.8.1
distlib==0.3.1
filelock==3.0.12
future==0.16.0
google-images-download==2.4.2
hgtk==0.1.3
idna==2.7
importlib-metadata==3.4.0
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
macholib==1.11
mccabe==0.6.1
notion==0.0.28
packaging==20.9
pdfminer==20191125
pefile==2018.8.8
Pillow==7.1.0
pip-review==1.1.0
py-wallpaper==0.0.1
pycryptodome==3.10.1
PyInstaller==3.6
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.12
python-docx==0.8.7
python-pptx==0.6.13
python-slugify==4.0.1
pytz==2021.1
pywin32==223
pywin32-ctypes==0.2.0
QDarkStyle==2.6.5
requests==2.20.0
schedule==0.6.0
selenium==3.14.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.24.2
virtualenv==20.4.0
virtualenvwrapper-win==1.2.6
webdriver-manager==3.4.2
wget==3.2
wrapt==1.12.1
XlsxWriter==1.1.0
zipp==3.4.0
